Revert "dieterブランチの追加"
