+++ src/App.js
@@ -1,0 +1,12 @@
+import { Router } from "./Router/Router";
+import "./scss/all.scss";
+
+function App() {
+  return (
+    <div>
+      <Router />
+    </div>
+  );
+}
+
+export default App;
