--- src/culinaryMate/App.css
@@ -1,163 +1,0 @@
-/* App.css */
-.App {
-  text-align: center;
-  font-family: "Segoe UI", Tahoma, Geneva, Verdana, sans-serif;
-  background-color: #ffebcd;
-  min-height: 100vh;
-  padding: 20px;
-}
-
-.App-header {
-  background-color: #ff1493;
-  padding: 20px;
-  color: white;
-  border-radius: 10px;
-  margin-bottom: 20px;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.2);
-}
-
-.App-title {
-  margin: 0;
-  font-size: 3em;
-  font-weight: bold;
-}
-
-.App-content {
-  background-color: #ffffff;
-  border-radius: 10px;
-  padding: 30px;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-}
-
-.App-button {
-  background-color: #ff69b4;
-  color: white;
-  border: none;
-  padding: 12px 24px;
-  text-align: center;
-  text-decoration: none;
-  display: inline-block;
-  font-size: 18px;
-  margin: 6px 4px;
-  cursor: pointer;
-  border-radius: 8px;
-  transition: background-color 0.3s, transform 0.3s;
-}
-
-.App-button:hover {
-  background-color: #ff1493;
-  transform: scale(1.05);
-}
-
-.App-input {
-  width: 100%;
-  padding: 14px 20px;
-  margin: 10px 0;
-  display: inline-block;
-  border: 2px solid #ff69b4;
-  border-radius: 8px;
-  box-sizing: border-box;
-}
-
-.App-table {
-  width: 100%;
-  border-collapse: collapse;
-  margin-top: 20px;
-}
-
-.App-table th,
-.App-table td {
-  border: 1px solid #ddd;
-  padding: 14px;
-  text-align: left;
-}
-
-.App-table th {
-  background-color: #ff69b4;
-  color: white;
-}
-
-.App-table tr:nth-child(even) {
-  background-color: #ffe4e1;
-}
-
-.App-table tr:hover {
-  background-color: #ffe4b5;
-}
-
-.response {
-  max-height: 300px;
-  overflow-y: auto;
-  background-color: #f9f9f9;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  padding: 10px;
-}
-.response-container {
-  margin-top: 20px;
-  background-color: #f9f9f9;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  padding: 10px;
-}
-
-.response-content {
-  max-height: 300px;
-  overflow-y: auto;
-  padding: 10px;
-  background-color: #ffffff;
-  border: 1px solid #e0e0e0;
-  border-radius: 4px;
-}
-
-.error-message {
-  color: red;
-  margin-top: 10px;
-}
-
-.response-container {
-  margin-top: 20px;
-  background-color: #f9f9f9;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  padding: 10px;
-}
-
-.response-content {
-  max-height: none;
-  overflow-y: visible;
-  padding: 10px;
-  background-color: #ffffff;
-  border: 1px solid #e0e0e0;
-  border-radius: 4px;
-  white-space: pre-wrap;
-  word-wrap: break-word;
-}
-
-.error-message {
-  color: red;
-  margin-top: 10px;
-}
-.response-container {
-  margin-top: 20px;
-  background-color: #f9f9f9;
-  border: 1px solid #ddd;
-  border-radius: 8px;
-  padding: 10px;
-}
-
-.response-content {
-  max-height: none;
-  overflow-y: visible;
-  padding: 10px;
-  background-color: #ffffff;
-  border: 1px solid #e0e0e0;
-  border-radius: 4px;
-  white-space: pre-wrap;
-  word-wrap: break-word;
-}
-
-.error-message {
-  color: red;
-  margin-top: 10px;
-}
