--- src/dieter/AppDieter.js
@@ -1,12 +1,0 @@
-import { Router } from "../Router/Router";
-import "./scss/all.scss";
-
-function AppDieter() {
-  return (
-    <div>
-      <Router />
-    </div>
-  );
-}
-
-export default AppDieter;
