+++ src/pages/Yotei.jsx
@@ -4,25 +4,20 @@ import interactionPlugin from "@fullcalendar/interaction";
 import jaLocale from "@fullcalendar/core/locales/ja";
 import { useNavigate } from "react-router-dom";
 import React from "react";
+import { supabase } from "../supabase";
 
 export const Yotei = (info) => {
   //認証情報とuserIdの取得
   const auth = info["auth"];
-  const user_id = info["user_id"];
-  const user_name = info["user_name"];
-  const workOut = info["workOut"];
-  //console.log(auth);
+  const userId = info["userId"];
+  const userName = info["userName"];
   const nav = useNavigate();
 
   const handleEventSecrect = (date) => {
     nav(`/dieter/Set/${date}`, {
-      state: {
-        id: user_id,
-        date: date,
-        user_name: user_name,
-        workOut: workOut,
-      },
+      state: { id: userId, date: date, userName: userName },
     });
+    // console.log(date);
   };
 
   return (
@@ -35,6 +30,7 @@ export const Yotei = (info) => {
           locale="ja"
           selectable={true}
           dateClick={(e) => handleEventSecrect(e.dateStr)}
+          // className="calender"
         />
       ) : (
         <FullCalendar
@@ -43,6 +39,7 @@ export const Yotei = (info) => {
           locales={[jaLocale]} // 追加
           locale="ja"
           selectable={true}
+          // className="calender"
         />
       )}
     </div>
