+++ src/scss/detaDetail.scss
